+++ Common/src/main/java/com/christofmeg/justenoughbreeding/jei/recipe/TemperRecipe.java
@@ -1,0 +1,3 @@
+package com.christofmeg.justenoughbreeding.jei.recipe;
+
+public record TemperRecipe(String temperIngredient, int temperValue) {}