+++ Forge/src/main/java/com/christofmeg/justenoughbreeding/jei/BreedingCategory.java
@@ -1,8 +1,8 @@
 package com.christofmeg.justenoughbreeding.jei;
 
 import com.christofmeg.justenoughbreeding.CommonConstants;
-import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.math.Axis;
+import com.christofmeg.justenoughbreeding.recipe.BreedingRecipe;
+import com.christofmeg.justenoughbreeding.utils.Utils;
 import mezz.jei.api.constants.VanillaTypes;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
@@ -15,39 +15,16 @@ import mezz.jei.api.recipe.category.IRecipeCategory;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.TamableAnimal;
-import net.minecraft.world.entity.animal.Cat;
-import net.minecraft.world.entity.animal.Fox;
-import net.minecraft.world.entity.animal.Ocelot;
-import net.minecraft.world.entity.animal.Panda;
-import net.minecraft.world.entity.animal.Pig;
-import net.minecraft.world.entity.animal.Turtle;
-import net.minecraft.world.entity.animal.Wolf;
-import net.minecraft.world.entity.animal.axolotl.Axolotl;
-import net.minecraft.world.entity.animal.frog.Frog;
-import net.minecraft.world.entity.animal.horse.Horse;
-import net.minecraft.world.entity.animal.sniffer.Sniffer;
-import net.minecraft.world.entity.monster.hoglin.Hoglin;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.AABB;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.joml.Quaternionf;
 
-public class BreedingCategory implements IRecipeCategory<BreedingCategory.BreedingRecipe> {
-
-    static final int ENTITY_CREATION_INTERVAL = 3000;
-    static final int ENTITY_RENDER_DISTANCE = 15728880;
+public class BreedingCategory implements IRecipeCategory<BreedingRecipe> {
 
     public static final RecipeType<BreedingRecipe> TYPE = new RecipeType<>(
             new ResourceLocation(CommonConstants.MOD_ID, "breeding"), BreedingRecipe.class);
@@ -208,157 +185,12 @@ public class BreedingCategory implements IRecipeCategory<BreedingCategory.Breedi
                 stack.drawString(font, abbreviatedEntityName, 0, 0, DyeColor.BLACK.getTextColor(), false);
             }
 
-            recipe.doRendering(stack.pose(), mouseX);
-
-        }
-    }
-
-    private static void renderEntity(@NotNull PoseStack stack, double mouseX, LivingEntity currentLivingEntity) {
-        // Set the desired position of the entity on the screen
-        int entityPosX = 31;
-        int entityPosY = 89;
-
-        float yaw = (float) (60 - mouseX); // Calculate the yaw based on the mouse position
-
-        stack.pushPose(); // Push the current pose onto the stack
-        stack.translate((float) entityPosX, (float) entityPosY, 50f); // Translate the entity's position
-
-        // Calculate the scaling factor based on the bounding box's largest dimension
-        AABB boundingBox = currentLivingEntity.getBoundingBox();
-        double largestDimension = Math.max(boundingBox.getXsize(), Math.max(boundingBox.getYsize(), boundingBox.getZsize()));
-
-        float desiredWidth = 30.0F;
-        float desiredHeight = 40.0F;
-
-        // Calculate the scaling factors for width and height
-        float scaleX = desiredWidth / (float) largestDimension;
-        float scaleY = desiredHeight / (float) largestDimension;
-
-        // Use the smaller of the two scaling factors to ensure the entity fits within the area
-        float scalingFactor = Math.min(scaleX, scaleY);
-
-        if (currentLivingEntity instanceof Frog) {
-            scalingFactor = 50;
-        }
-
-        if (currentLivingEntity instanceof Axolotl || currentLivingEntity instanceof Cat ||
-                currentLivingEntity instanceof Pig || currentLivingEntity instanceof Wolf) {
-            scalingFactor = 25;
-        }
-
-        if (currentLivingEntity instanceof Ocelot || currentLivingEntity instanceof Fox
-                || currentLivingEntity instanceof Turtle) {
-            scalingFactor = 20;
-        }
-
-        if (currentLivingEntity instanceof Hoglin || currentLivingEntity instanceof Horse
-                || currentLivingEntity instanceof Panda) {
-            scalingFactor = 15;
-        }
-
-        if (currentLivingEntity instanceof Sniffer) {
-            scalingFactor = 10;
-        }
-
-        stack.scale(scalingFactor, scalingFactor, scalingFactor); // Scale the entity to fit within the desired area
-        stack.mulPose(Axis.ZP.rotationDegrees(180.0F)); // Rotate the entity to face a certain direction
-
-        float yawRadians = -(yaw / 40.F) * 20.0F; // Calculate the yaw angle in radians for the entity's rotation
-
-        // Apply the calculated yaw angle to the entity's rotation properties
-        currentLivingEntity.yBodyRot = yawRadians;
-        currentLivingEntity.setYRot(yawRadians);
-        currentLivingEntity.yHeadRot = yawRadians;
-        currentLivingEntity.yHeadRotO = yawRadians;
-
-        stack.translate(0.0F, currentLivingEntity.getMyRidingOffset(currentLivingEntity), 0.0F); // Translate the entity vertically to adjust its position
-
-        Minecraft instance = Minecraft.getInstance();
-        EntityRenderDispatcher entityRenderDispatcher = instance.getEntityRenderDispatcher(); // Get the entity rendering dispatcher
-        entityRenderDispatcher.overrideCameraOrientation(new Quaternionf(0.0F, 0.0F, 0.0F, 1.0F)); // Override the camera orientation for rendering
-        entityRenderDispatcher.setRenderShadow(false); // Disable rendering shadows for the entity
-
-        // Get the buffer source for rendering
-        final MultiBufferSource.BufferSource bufferSource = instance.renderBuffers().bufferSource();
-
-        // Render the currentLivingEntity using the entityRenderDispatcher
-        entityRenderDispatcher.render(currentLivingEntity, 0.0D, 0.0D, 0.0D, 0.0F, 1.0F, stack, bufferSource, ENTITY_RENDER_DISTANCE);
-
-        bufferSource.endBatch(); // End the rendering batch
-        entityRenderDispatcher.setRenderShadow(true); // Re-enable rendering shadows
-
-        stack.popPose(); // Pop the pose from the stack to revert transformations
-    }
-
-    public static class BreedingRecipe {
-        private LivingEntity currentLivingEntity = null;
-        private long lastEntityCreationTime = 0;
-
-        private final EntityType<?> entityType;
-        private final Ingredient breedingCatalyst;
-        private final ItemStack spawnEgg;
-        @Nullable
-        private final Boolean needsToBeTamed;
-        private final Ingredient resultItemStack;
-        @Nullable
-        private final Ingredient extraInputStack;
-        @Nullable
-        private final Boolean animalTrusting;
-
-        public BreedingRecipe(EntityType<?> entityType, Ingredient breedingCatalyst, ItemStack spawnEgg, @Nullable Boolean needsToBeTamed, @Nullable Ingredient resultItemStack, @Nullable Ingredient extraInputStack, @Nullable Boolean animalTrusting) {
-            this.entityType = entityType;
-            this.breedingCatalyst = breedingCatalyst;
-            this.spawnEgg = spawnEgg;
-            this.needsToBeTamed = needsToBeTamed;
-            this.resultItemStack = resultItemStack;
-            this.extraInputStack = extraInputStack;
-            this.animalTrusting = animalTrusting;
-        }
-
-        private void doRendering(PoseStack stack, double mouseX) {
-            long currentTime = System.currentTimeMillis();
-            Level level = Minecraft.getInstance().level;
-
-            if (level != null && (currentLivingEntity == null || currentTime - lastEntityCreationTime >= ENTITY_CREATION_INTERVAL)) {
-                currentLivingEntity = (LivingEntity) entityType.create(level);
-                lastEntityCreationTime = currentTime;
-            }
-
+            LivingEntity currentLivingEntity = recipe.doRendering();
             if (currentLivingEntity != null) {
-                if (currentLivingEntity instanceof TamableAnimal tamableAnimal) {
-                    tamableAnimal.setTame(true);
-                }
-                renderEntity(stack, mouseX, currentLivingEntity);
+                Utils.renderEntity(stack.pose(), mouseX, currentLivingEntity);
             }
+
         }
-
     }
-
-    //TODO https://www.curseforge.com/minecraft/mc-mods/betteranimalsplus
-    //TODO https://www.curseforge.com/minecraft/mc-mods/upgrade-aquatic
-    //TODO https://www.curseforge.com/minecraft/mc-mods/galosphere
-    //TODO https://www.curseforge.com/minecraft/mc-mods/buzzier-bees
-    //TODO https://www.curseforge.com/minecraft/mc-mods/exotic-birds
-    //TODO https://www.curseforge.com/minecraft/mc-mods/creatures-and-beasts
-    //TODO https://www.curseforge.com/minecraft/mc-mods/extended-mushrooms
-    //TODO https://www.curseforge.com/minecraft/mc-mods/more-babies
-    //TODO https://www.curseforge.com/minecraft/mc-mods/goodall
-    //TODO https://www.curseforge.com/minecraft/mc-mods/energeticsheep
-    //TODO https://www.curseforge.com/minecraft/mc-mods/feywild
-    //TODO https://www.curseforge.com/minecraft/mc-mods/earth2java
-    //TODO https://www.curseforge.com/minecraft/mc-mods/unusual-end
-    //TODO https://www.curseforge.com/minecraft/mc-mods/vanilla-degus
-    //TODO https://www.curseforge.com/minecraft/mc-mods/fins-and-tails
-    //TODO https://www.curseforge.com/minecraft/mc-mods/realistic-horse-genetics
-    //TODO https://www.curseforge.com/minecraft/mc-mods/critters-and-companions
-    //TODO https://www.curseforge.com/minecraft/mc-mods/the-undergarden
-    //TODO https://www.curseforge.com/minecraft/mc-mods/productivebees
-    //TODO https://www.curseforge.com/minecraft/mc-mods/roost-ultimate
-    //TODO https://www.curseforge.com/minecraft/mc-mods/ender-zoology
-    //TODO https://www.curseforge.com/minecraft/mc-mods/primal-reservation
-    //TODO https://www.curseforge.com/minecraft/mc-mods/baby-fat
-    //TODO https://www.curseforge.com/minecraft/mc-mods/fins-and-tails
-    //TODO https://www.curseforge.com/minecraft/mc-mods/minecraft-earth-mod
-
 
 }