+++ NeoForge/src/main/java/com/christofmeg/justenoughbreeding/config/integrated/AetherReduxIntegration.java
@@ -1,0 +1,28 @@
+package com.christofmeg.justenoughbreeding.config.integrated;
+
+import com.christofmeg.justenoughbreeding.CommonStrings;
+import com.christofmeg.justenoughbreeding.utils.CommonUtils;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class AetherReduxIntegration {
+
+    static final String MOD = "aether_redux";
+
+    static final List<String> animalNames = new ArrayList<>();
+    static final Map<String, String> ingredients = new HashMap<>();
+    static final Map<String, String> resultEggs = new HashMap<>();
+    static final Map<String, Integer> eggsAmountMin = new HashMap<>();
+    static final Map<String, Integer> eggsAmountMax = new HashMap<>();
+    static final Map<String, Integer> breedingCooldown = new HashMap<>();
+
+    public static void init() {
+        CommonUtils.addAnimal("mykapod", CommonStrings.MYKAPOD_TEMPTATION_ITEMS_TAG, animalNames, ingredients, breedingCooldown);
+        CommonUtils.addAnimal("shimmercow", CommonStrings.SHIMMERCOW_TEMPTATION_ITEMS_TAG, animalNames, ingredients, breedingCooldown);
+        CommonUtils.addAnimalNames(animalNames, ingredients, MOD, breedingCooldown, resultEggs, eggsAmountMin, eggsAmountMax);
+    }
+
+}