+++ NeoForge/src/main/java/com/christofmeg/justenoughbreeding/config/integrated/SophisticatedWolvesIntegration.java
@@ -1,0 +1,31 @@
+package com.christofmeg.justenoughbreeding.config.integrated;
+
+import com.christofmeg.justenoughbreeding.CommonStrings;
+import com.christofmeg.justenoughbreeding.utils.CommonUtils;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class SophisticatedWolvesIntegration {
+
+    static final String MOD = "sophisticated_wolves";
+
+    static final List<String> animalNames = new ArrayList<>();
+    static final List<String> tamableOnly = new ArrayList<>();
+    static final Map<String, String> ingredients = new HashMap<>();
+    static final Map<String, Boolean> needsToBeTamed = new HashMap<>();
+    static final Map<String, Integer> breedingCooldown = new HashMap<>();
+    static final Map<String, String> tamingIngredients = new HashMap<>();
+    static final Map<String, Integer> tamingChance = new HashMap<>();
+
+    public static void init() {
+        CommonUtils.addAnimalWithTamedTag("sophisticated_wolf", CommonStrings.DOG_TREAT, animalNames, ingredients, breedingCooldown, needsToBeTamed);
+        CommonUtils.addAnimalNames(animalNames, ingredients, MOD, breedingCooldown, needsToBeTamed, "dog_egg");
+
+        CommonUtils.addTamableOnly("sophisticated_wolf", CommonStrings.BONE, tamableOnly, tamingIngredients, tamingChance);
+        CommonUtils.addTamableAnimalNames(tamableOnly, tamingIngredients, tamingChance, MOD, "dog_egg");
+    }
+
+}