+++ NeoForge/src/main/java/com/christofmeg/justenoughbreeding/utils/Utils.java
@@ -1,22 +1,34 @@
 package com.christofmeg.justenoughbreeding.utils;
 
-import com.christofmeg.justenoughbreeding.CommonConstants;
 import com.christofmeg.justenoughbreeding.JustEnoughBreeding;
-import com.christofmeg.justenoughbreeding.jei.BreedingCategory;
-import mezz.jei.api.registration.IRecipeRegistration;
+import com.christofmeg.justenoughbreeding.recipe.BreedingRecipe;
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.math.Axis;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.entity.EntityRenderDispatcher;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.animal.*;
+import net.minecraft.world.entity.animal.axolotl.Axolotl;
+import net.minecraft.world.entity.animal.frog.Frog;
+import net.minecraft.world.entity.animal.horse.Horse;
+import net.minecraft.world.entity.animal.sniffer.Sniffer;
+import net.minecraft.world.entity.monster.hoglin.Hoglin;
 import net.minecraft.world.food.FoodProperties;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraft.world.phys.AABB;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.joml.Quaternionf;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 public class Utils {
@@ -24,19 +36,17 @@ public class Utils {
     public static String getEdibleMeatItemNames(boolean includeRottenFlesh) {
         List<String> edibleMeatItemNames = new ArrayList<>();
 
-        for (ResourceLocation key : ForgeRegistries.ITEMS.getKeys()) {
-            Item item = ForgeRegistries.ITEMS.getValue(key);
-            if (item != null) {
-                FoodProperties foodProperties = item.getFoodProperties(item.getDefaultInstance(), null);
-                if(includeRottenFlesh) {
-                    if (foodProperties != null && item.isEdible() && foodProperties.isMeat()) {
-                        edibleMeatItemNames.add(key.toString());
-                    }
+        for (ResourceLocation key : BuiltInRegistries.ITEM.keySet()) {
+            Item item = BuiltInRegistries.ITEM.get(key);
+            FoodProperties foodProperties = item.getFoodProperties(item.getDefaultInstance(), null);
+            if(includeRottenFlesh) {
+                if (foodProperties != null && item.isEdible() && foodProperties.isMeat()) {
+                    edibleMeatItemNames.add(key.toString());
                 }
-                else {
-                    if (foodProperties != null && item.isEdible() && foodProperties.isMeat() && item != Items.ROTTEN_FLESH) {
-                        edibleMeatItemNames.add(key.toString());
-                    }
+            }
+            else {
+                if (foodProperties != null && item.isEdible() && foodProperties.isMeat() && item != Items.ROTTEN_FLESH) {
+                    edibleMeatItemNames.add(key.toString());
                 }
             }
         }
@@ -44,65 +54,15 @@ public class Utils {
         return String.join(", ", edibleMeatItemNames);
     }
 
-    public static String getAllItems() {
-        List<String> allItems = new ArrayList<>();
-
-        for (ResourceLocation key : ForgeRegistries.ITEMS.getKeys()) {
-            Item item = ForgeRegistries.ITEMS.getValue(key);
-            if (item != null) {
-                allItems.add(key.toString());
-            }
-        }
-
-        return String.join(", ", allItems);
-    }
-
-    public static void registerMobBreedingRecipes(IRecipeRegistration registration) {
-        List<String> sortedMobNames = new ArrayList<>(CommonConstants.ingredientConfigs.keySet());
-        Collections.sort(sortedMobNames);
-
-        for (String mobName : sortedMobNames) {
-            if (mobName != null) {
-                if (CommonConstants.ingredientConfigs != null && CommonConstants.spawnEggConfigs != null) {
-                    if (CommonConstants.spawnEggConfigs.get(mobName) != null) {
-                        String mobIngredients = CommonConstants.ingredientConfigs.get(mobName).get();
-                        String mobSpawnEgg = CommonConstants.spawnEggConfigs.get(mobName).get();
-                        String mobResultItem = CommonConstants.eggResultConfigs.get(mobName) != null ? CommonConstants.eggResultConfigs.get(mobName).get() : "";
-                        int mobMinResultCount = CommonConstants.eggMinAmountConfigs.get(mobName) != null ? CommonConstants.eggMinAmountConfigs.get(mobName).get() : 1;
-                        int mobMaxResultCount = CommonConstants.eggMaxAmountConfigs.get(mobName) != null ? CommonConstants.eggMaxAmountConfigs.get(mobName).get() : 1;
-
-                        if (mobIngredients != null && mobSpawnEgg != null) {
-                            Ingredient combinedIngredient = createCombinedIngredient(mobIngredients);
-                            List<Ingredient> combinedResultIngredient = Utils.createCombinedResultIngredients(mobResultItem, mobMinResultCount, mobMaxResultCount);
-                            Item spawnEggItem = JustEnoughBreeding.getItemFromLoaderRegistries(new ResourceLocation(mobSpawnEgg.trim()));
-
-                            if (spawnEggItem instanceof SpawnEggItem spawnEgg) {
-                                EntityType<?> entityType = spawnEgg.getType(null);
-                                Boolean needsToBeTamed = CommonConstants.animalTamedConfigs.get(mobName);
-                                Boolean animalTrusting = CommonConstants.animalTrustingConfigs.get(mobName);
-                                BreedingCategory.BreedingRecipe breedingRecipe = createBreedingRecipe(entityType, combinedIngredient, spawnEggItem, needsToBeTamed, combinedResultIngredient, animalTrusting);
-
-                                registration.addRecipes(BreedingCategory.TYPE, Collections.singletonList(breedingRecipe));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private static List<Ingredient> createCombinedResultIngredients(String mobIngredients, int minCount, int maxCount) {
+    public static List<Ingredient> createCombinedResultIngredients(String mobIngredients, int minCount, int maxCount) {
         String[] ingredientIds = mobIngredients.split(",");
         List<Ingredient> resultIngredients = new ArrayList<>();
 
         List<ItemStack> combinedItemStacks = new ArrayList<>();
-
         for (int count = minCount; count <= maxCount; count++) {
             for (String ingredientId : ingredientIds) {
                 Item ingredientItem = JustEnoughBreeding.getItemFromLoaderRegistries(new ResourceLocation(ingredientId.trim()));
-                if (ingredientItem != null) {
-                    combinedItemStacks.add(new ItemStack(ingredientItem, count));
-                }
+                combinedItemStacks.add(new ItemStack(ingredientItem, count));
             }
         }
 
@@ -110,7 +70,7 @@ public class Utils {
         return resultIngredients;
     }
 
-    private static BreedingCategory.BreedingRecipe createBreedingRecipe(EntityType<?> entityType, Ingredient combinedIngredient, Item spawnEggItem, Boolean needsToBeTamed, List<Ingredient> resultItemStacks, Boolean animalTrusting) {
+    public static BreedingRecipe createBreedingRecipe(EntityType<?> entityType, Ingredient combinedIngredient, Item spawnEggItem, Boolean needsToBeTamed, List<Ingredient> resultItemStacks, Boolean animalTrusting, @Nullable Ingredient combinedExtraIngredient) {
         List<ItemStack> mergedResultItemStacks = new ArrayList<>();
 
         for (Ingredient resultItemStack : resultItemStacks) {
@@ -118,18 +78,18 @@ public class Utils {
             mergedResultItemStacks.addAll(Arrays.asList(stacks));
         }
 
-        return new BreedingCategory.BreedingRecipe(
+        return new BreedingRecipe(
                 entityType,
                 combinedIngredient,
                 new ItemStack(spawnEggItem),
                 needsToBeTamed,
                 Ingredient.of(mergedResultItemStacks.toArray(new ItemStack[0])),
-                null,
+                combinedExtraIngredient,
                 animalTrusting
         );
     }
 
-    private static Ingredient createCombinedIngredient(String mobIngredients) {
+    public static Ingredient createCombinedIngredient(String mobIngredients) {
         String[] ingredientIds = mobIngredients.split(",");
         List<Ingredient> combinedIngredients = new ArrayList<>();
 
@@ -138,13 +98,91 @@ public class Utils {
                 combinedIngredients.add(CommonUtils.createTagIngredient(ingredientId));
             } else {
                 Item ingredientItem = JustEnoughBreeding.getItemFromLoaderRegistries(new ResourceLocation(ingredientId.trim()));
-                if (ingredientItem != null) {
-                    combinedIngredients.add(Ingredient.of(new ItemStack(ingredientItem)));
-                }
+                combinedIngredients.add(Ingredient.of(new ItemStack(ingredientItem)));
             }
         }
 
-        return Ingredient.merge(combinedIngredients);
+        return Ingredient.of(Arrays.stream(combinedIngredients.toArray(Ingredient[]::new))
+                .flatMap(ingredient -> Arrays.stream(ingredient.getItems()))
+                .distinct()
+                .toArray(ItemStack[]::new));
+    }
+    public static void renderEntity(@NotNull PoseStack stack, double mouseX, LivingEntity currentLivingEntity) {
+        // Set the desired position of the entity on the screen
+        int entityPosX = 31;
+        int entityPosY = 89;
+        int ENTITY_RENDER_DISTANCE = 15728880;
+
+        float yaw = (float) (60 - mouseX); // Calculate the yaw based on the mouse position
+
+        stack.pushPose(); // Push the current pose onto the stack
+        stack.translate((float) entityPosX, (float) entityPosY, 50f); // Translate the entity's position
+
+        // Calculate the scaling factor based on the bounding box's largest dimension
+        AABB boundingBox = currentLivingEntity.getBoundingBox();
+        double largestDimension = Math.max(boundingBox.getXsize(), Math.max(boundingBox.getYsize(), boundingBox.getZsize()));
+
+        float desiredWidth = 30.0F;
+        float desiredHeight = 40.0F;
+
+        // Calculate the scaling factors for width and height
+        float scaleX = desiredWidth / (float) largestDimension;
+        float scaleY = desiredHeight / (float) largestDimension;
+
+        // Use the smaller of the two scaling factors to ensure the entity fits within the area
+        float scalingFactor = Math.min(scaleX, scaleY);
+
+        if (currentLivingEntity instanceof Frog) {
+            scalingFactor = 50;
+        }
+
+        if (currentLivingEntity instanceof Axolotl || currentLivingEntity instanceof Cat ||
+                currentLivingEntity instanceof Pig || currentLivingEntity instanceof Wolf) {
+            scalingFactor = 25;
+        }
+
+        if (currentLivingEntity instanceof Ocelot || currentLivingEntity instanceof Fox
+                || currentLivingEntity instanceof Turtle) {
+            scalingFactor = 20;
+        }
+
+        if (currentLivingEntity instanceof Hoglin || currentLivingEntity instanceof Horse
+                || currentLivingEntity instanceof Panda) {
+            scalingFactor = 15;
+        }
+
+        if (currentLivingEntity instanceof Sniffer) {
+            scalingFactor = 10;
+        }
+
+        stack.scale(scalingFactor, scalingFactor, scalingFactor); // Scale the entity to fit within the desired area
+        stack.mulPose(Axis.ZP.rotationDegrees(180.0F)); // Rotate the entity to face a certain direction
+
+        float yawRadians = -(yaw / 40.F) * 20.0F; // Calculate the yaw angle in radians for the entity's rotation
+
+        // Apply the calculated yaw angle to the entity's rotation properties
+        currentLivingEntity.yBodyRot = yawRadians;
+        currentLivingEntity.setYRot(yawRadians);
+        currentLivingEntity.yHeadRot = yawRadians;
+        currentLivingEntity.yHeadRotO = yawRadians;
+
+        stack.translate(0.0F, currentLivingEntity.getMyRidingOffset(currentLivingEntity), 0.0F); // Translate the entity vertically to adjust its position
+
+        Minecraft instance = Minecraft.getInstance();
+        EntityRenderDispatcher entityRenderDispatcher = instance.getEntityRenderDispatcher(); // Get the entity rendering dispatcher
+        entityRenderDispatcher.overrideCameraOrientation(new Quaternionf(0.0F, 0.0F, 0.0F, 1.0F)); // Override the camera orientation for rendering
+        entityRenderDispatcher.setRenderShadow(false); // Disable rendering shadows for the entity
+
+        // Get the buffer source for rendering
+        final MultiBufferSource.BufferSource bufferSource = instance.renderBuffers().bufferSource();
+
+        // Render the currentLivingEntity using the entityRenderDispatcher
+        entityRenderDispatcher.render(currentLivingEntity, 0.0D, 0.0D, 0.0D, 0.0F, 1.0F, stack, bufferSource, ENTITY_RENDER_DISTANCE);
+
+        bufferSource.endBatch(); // End the rendering batch
+        entityRenderDispatcher.setRenderShadow(true); // Re-enable rendering shadows
+
+        stack.popPose(); // Pop the pose from the stack to revert transformations
     }
 
 }