+++ Quilt/src/main/java/com/christofmeg/justenoughbreeding/JustEnoughBreeding.java
@@ -1,28 +1,26 @@
 package com.christofmeg.justenoughbreeding;
 
-import com.christofmeg.justenoughbreeding.config.JEBConfig;
-import fuzs.forgeconfigapiport.api.config.v2.ForgeConfigRegistry;
+import com.christofmeg.justenoughbreeding.config.JEBIntegration;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Item;
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.fml.config.ModConfig;
 import org.quiltmc.loader.api.ModContainer;
 import org.quiltmc.qsl.base.api.entrypoint.client.ClientModInitializer;
 
 public class JustEnoughBreeding implements ClientModInitializer {
 
-    private static final ForgeConfigSpec.Builder BUILDER = new ForgeConfigSpec.Builder();
-    @SuppressWarnings("unused")
-    public static final JEBConfig GENERAL = new JEBConfig(BUILDER);
-
     @Override
     public void onInitializeClient(ModContainer mod) {
-        ForgeConfigRegistry.INSTANCE.register(CommonConstants.MOD_ID, ModConfig.Type.CLIENT, BUILDER.build());
+        JEBIntegration.init();
     }
 
     public static Item getItemFromLoaderRegistries(ResourceLocation resourceLocation) {
         return BuiltInRegistries.ITEM.get(resourceLocation);
+    }
+
+    public static EntityType<?> getEntityFromLoaderRegistries(ResourceLocation resourceLocation) {
+        return BuiltInRegistries.ENTITY_TYPE.get(resourceLocation);
     }
 
 }