+++ Forge/changelog.md
@@ -1,8 +1,38 @@
-1.0.10:
-- Integration to Biome Makeover
-- Added ukrainian translation by [MeDustyy](https://github.com/MeDustyy)
-- More background rework to integrations
+1.2.1:
+- Call Of The Wild Integration
+- Creatures And Beasts Integration
+- Winter Overhaul Integration
+- Critters And Companions Integration
 
-1.0.11:
-- Added some missing entries to Naturalist integration
-- Did more background work with integrations
+1.2.0:
+- Added support for EMI
+
+1.1.1:
+- Added DisplayTest and DistExecutor
+- Fix [#12](#12)
+
+1.1.0:
+- Added support for REI
+
+1.0.16:
+- Actually enable ChocoCraft integration
+
+1.0.15:
+- Ice and Fire Integration
+- Aether Integration
+- Aether: Redux Integration
+- Deep Aether Integration
+- Implements [#11](#11)
+
+1.0.14:
+- Added Quark Parrot Eggs
+- ChocoCraft integration
+
+1.0.13:
+- Added ja_jp.json by [EndilCrafter](https://github.com/EndilCrafter)
+- Move away from using config file.
+- Removed dependency to Forge Config API Port
+
+1.0.12:
+- Integration to Alex's Mobs
+- More background work on integrations